Relocate simulation config and align trace setup
